# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MEI Fácil API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  cpf: string;
  cnpj: string;
  /** @nullable */
  phone?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RefreshInput {
  refreshToken: string;
}

export interface ResetPasswordInput {
  email: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  cpf: string;
  cnpj: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  openingDate?: string | null;
  /** @nullable */
  activity?: string | null;
  /** @nullable */
  category?: string | null;
  annualLimit?: number;
  onboardingCompleted?: boolean;
  createdAt?: string;
}

export interface AuthResponse {
  accessToken: string;
  refreshToken: string;
  user: UserProfile;
}

export interface ProfileUpdate {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  openingDate?: string | null;
  /** @nullable */
  activity?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  annualLimit?: number | null;
  /** @nullable */
  onboardingCompleted?: boolean | null;
}

export type DashboardSummaryStatus = typeof DashboardSummaryStatus[keyof typeof DashboardSummaryStatus];


export const DashboardSummaryStatus = {
  regular: 'regular',
  attention: 'attention',
  risk: 'risk',
} as const;

export type TaxPaymentStatus = typeof TaxPaymentStatus[keyof typeof TaxPaymentStatus];


export const TaxPaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  overdue: 'overdue',
} as const;

export interface TaxPayment {
  id: number;
  userId: number;
  competence: string;
  dueDate: string;
  amount: number;
  status: TaxPaymentStatus;
  /** @nullable */
  paidAt?: string | null;
  /** @nullable */
  daysUntilDue?: number | null;
  createdAt: string;
}

export interface DashboardSummary {
  annualRevenue: number;
  annualLimit: number;
  usedPercent: number;
  status: DashboardSummaryStatus;
  nextDueTax: TaxPayment;
  monthlyRevenue: number;
  unreadAlerts: number;
  projectedAnnual?: number;
}

export interface Revenue {
  id: number;
  userId: number;
  amount: number;
  date: string;
  category: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface RevenueInput {
  amount: number;
  date: string;
  category: string;
  /** @nullable */
  description?: string | null;
}

export interface RevenueUpdate {
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface RevenueListResponse {
  entries: Revenue[];
  total: number;
}

export type RevenueStatsAlertLevel = typeof RevenueStatsAlertLevel[keyof typeof RevenueStatsAlertLevel];


export const RevenueStatsAlertLevel = {
  safe: 'safe',
  warning: 'warning',
  danger: 'danger',
  exceeded: 'exceeded',
} as const;

export interface RevenueStats {
  totalAnnual: number;
  totalMonthly: number;
  avgMonthly: number;
  projectedAnnual: number;
  usedPercent: number;
  annualLimit: number;
  monthsWithData: number;
  alertLevel?: RevenueStatsAlertLevel;
}

export interface MonthlyRevenue {
  month: number;
  year: number;
  total: number;
  label?: string;
}

export interface TaxInput {
  competence: string;
  dueDate: string;
  amount: number;
}

export interface TaxPaymentUpdate {
  /** @nullable */
  paidAt?: string | null;
}

export interface PenaltySimulationInput {
  amount: number;
  daysLate: number;
}

export interface PenaltySimulation {
  originalAmount: number;
  fine: number;
  interest: number;
  total: number;
  daysLate: number;
}

export type DocumentFolder = typeof DocumentFolder[keyof typeof DocumentFolder];


export const DocumentFolder = {
  das: 'das',
  notas_fiscais: 'notas_fiscais',
  declaracoes: 'declaracoes',
  certidoes: 'certidoes',
  outros: 'outros',
} as const;

export interface Document {
  id: number;
  userId: number;
  name: string;
  folder: DocumentFolder;
  fileType: string;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface DocumentInput {
  name: string;
  folder: string;
  fileType: string;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  description?: string | null;
}

export type AlertType = typeof AlertType[keyof typeof AlertType];


export const AlertType = {
  vencimento: 'vencimento',
  faturamento: 'faturamento',
  documento: 'documento',
  regularizacao: 'regularizacao',
} as const;

export type AlertPriority = typeof AlertPriority[keyof typeof AlertPriority];


export const AlertPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Alert {
  id: number;
  userId: number;
  type: AlertType;
  title: string;
  description: string;
  priority: AlertPriority;
  isRead: boolean;
  createdAt: string;
}

export type DeclarationStatus = typeof DeclarationStatus[keyof typeof DeclarationStatus];


export const DeclarationStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export interface Declaration {
  year: number;
  annualRevenue: number;
  monthsRegistered: number;
  status: DeclarationStatus;
  completionPercent?: number;
}

export interface ChecklistItem {
  id: number;
  userId: number;
  year: number;
  label: string;
  completed: boolean;
}

export interface ChecklistUpdate {
  completed: boolean;
}

export interface AiMessageInput {
  message: string;
}

export type AiMessageRole = typeof AiMessageRole[keyof typeof AiMessageRole];


export const AiMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface AiMessage {
  id: number;
  userId: number;
  role: AiMessageRole;
  content: string;
  createdAt: string;
}

export type ListRevenueParams = {
/**
 * @nullable
 */
month?: number | null;
/**
 * @nullable
 */
year?: number | null;
/**
 * @nullable
 */
category?: string | null;
};

export type GetRevenueStatsParams = {
/**
 * @nullable
 */
year?: number | null;
};

export type GetMonthlyRevenueParams = {
/**
 * @nullable
 */
year?: number | null;
};

export type ListTaxesParams = {
/**
 * @nullable
 */
year?: number | null;
/**
 * @nullable
 */
status?: string | null;
};

export type ListDocumentsParams = {
/**
 * @nullable
 */
folder?: string | null;
/**
 * @nullable
 */
search?: string | null;
};

export type ListAlertsParams = {
/**
 * @nullable
 */
unreadOnly?: boolean | null;
};

export type GetDeclarationParams = {
/**
 * @nullable
 */
year?: number | null;
};

export type GetDeclarationChecklistParams = {
/**
 * @nullable
 */
year?: number | null;
};