# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MEI Fácil API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts
      security:
        - bearerAuth: []
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}/read:
    patch:
      operationId: markAlertRead
      tags: [alerts]
      summary: Mark alert as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /declaration:
    get:
      operationId: getDeclaration
      tags: [declaration]
      summary: Get annual declaration summary
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Declaration"
  /declaration/checklist:
    get:
      operationId: getDeclarationChecklist
      tags: [declaration]
      summary: Get declaration checklist items
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChecklistItem"
  /declaration/checklist/{id}:
    patch:
      operationId: updateChecklistItem
      tags: [declaration]
      summary: Update checklist item status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChecklistItem"
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Send a message to the AI assistant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiMessage"
  /ai/conversations:
    get:
      operationId: getAiConversations
      tags: [ai]
      summary: Get AI conversation history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiMessage"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterInput:
      type: object
      required: [name, email, password, cpf, cnpj]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        cpf:
          type: string
        cnpj:
          type: string
        phone:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    RefreshInput:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    ResetPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    AuthResponse:
      type: object
      required: [accessToken, refreshToken, user]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      required: [id, name, email, cpf, cnpj]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        cpf:
          type: string
        cnpj:
          type: string
        phone:
          type: ["string", "null"]
        openingDate:
          type: ["string", "null"]
        activity:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        annualLimit:
          type: number
        onboardingCompleted:
          type: boolean
        createdAt:
          type: string
...
    Alert:
      type: object
      required: [id, userId, type, title, description, priority, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [vencimento, faturamento, documento, regularizacao]
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, critical]
        isRead:
          type: boolean
        createdAt:
          type: string
    Declaration:
      type: object
      required: [year, annualRevenue, monthsRegistered, status]
      properties:
        year:
          type: integer
        annualRevenue:
          type: number
        monthsRegistered:
          type: integer
        status:
          type: string
          enum: [pending, in_progress, completed]
        completionPercent:
          type: number
    ChecklistItem:
      type: object
      required: [id, userId, year, label, completed]
      properties:
        id:
          type: integer
        userId:
          type: integer
        year:
          type: integer
        label:
          type: string
        completed:
          type: boolean
    ChecklistUpdate:
      type: object
      required: [completed]
      properties:
        completed:
          type: boolean
    AiMessageInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
    AiMessage:
      type: object
      required: [id, userId, role, content, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string